--- unit_tests/sources/declarative/decoders/parsers/__init__.py
@@ -1,3 +1,0 @@
-#
-# Copyright (c) 2024 Airbyte, Inc., all rights reserved.
-#
