+++ unit_tests/sources/declarative/decoders/test_composite_decoder.py
@@ -14,6 +14,7 @@ from airbyte_cdk.sources.declarative.decoders.composite_raw_decoder import (
     CsvParser,
     GzipParser,
     JsonLineParser,
+    JsonParser
 )
 
 
@@ -117,3 +118,28 @@ def test_composite_raw_decoder_jsonline_parser(requests_mock, encoding: str):
     for _ in composite_raw_decoder.decode(response):
         counter += 1
     assert counter == 3
+
+@pytest.mark.parametrize(
+    "raw_data, expected",
+    [
+        (json.dumps({"data-type": "string"}), {"data-type": "string"}),
+        (json.dumps({"data-type": "bytes"}).encode("utf-8"), {"data-type": "bytes"}),
+        (
+            bytearray(json.dumps({"data-type": "bytearray"}).encode("utf-8")),
+            {"data-type": "bytearray"},
+        ),
+        (json.dumps([{"id": 1}, {"id": 2}]), [{"id": 1}, {"id": 2}]),
+    ],
+    ids=[
+        "test_with_str",
+        "test_with_bytes",
+        "test_with_bytearray",
+        "test_with_string_data_containing_list",
+    ],
+)
+def test_json_parser_with_valid_data(raw_data, expected):
+    for i, actual in enumerate(JsonParser().parse(raw_data)):
+        if isinstance(expected, list):
+            assert actual == expected[i]
+        else:
+            assert actual == expected
